test: 에러 안나게 수정 (구조 변경에 맞게)
